+++ tests/wdio/raw-spec.js
@@ -1,7 +1,7 @@
 Terra.describeViewports('raw route', ['tiny', 'huge'], () => {
   it('checks accessibility on the raw route', () => {
     browser.url('/raw/secondary-nav-test/terra-dev-site/themed');
-    Terra.validates.element({ selector: '#root' });
+    Terra.validates.element('raw route', { selector: '#root' });
   });
 
   it('checks that the # route redirects ', () => {
@@ -11,7 +11,7 @@ Terra.describeViewports('raw route', ['tiny', 'huge'], () => {
 
   it('does not prevent navigation', () => {
     browser.url('/raw/test/terra-dev-site/navigation-prompt');
-    browser.click('[id="PendingStateButton"]');
+    $('[id="PendingStateButton"]').click();
     Terra.validates.element('navigation-prompt', { selector: '#root' });
     browser.url('/raw/test/terra-dev-site/relative-link');
     Terra.validates.element('redirected', { selector: '#root' });