+++ tests/wdio/singlePage-spec.js
@@ -1,6 +1,6 @@
 Terra.describeViewports('home', ['tiny', 'huge'], () => {
   it('checks accessibility', () => {
     browser.url('/single-page-test');
-    Terra.validates.element({ selector: '#root' });
+    Terra.validates.element('accessibility', { selector: '#root' });
   });
 });